Froze DeviceCompanyState refactoring into V2, broken (see description)
I was trying to move the actual Doctrine entities relating to DeviceCompanyState (DCS, Functionality etc.) into V2, but I encountered a blocking problem: DCS don't have a primary key column `id`, its PK is compsed of 3 FK columns (association table), which means that to work with Doctrine, as we need a getId, we're forced to add migrations.
As, per specification, we decided to not do any model change, we can't implement this right now.

This commit/rollback is only kept for historical purposes, in case we decide to migrate DCS into V2 under Doctrine.